sharpsl_pm: Fix offline charging.
The sharpsl_pm driver needs to read the AC_IN voltage
periodically while the device is suspended for
proper offline charging. The device that provides these voltage
readings is read through the pxa SPI bus which is unavailable while
the device is suspended.

This patch adds functions that enable raw reads of
these voltage values.

This is not a proper fix for this problem. We can
never be sure when the SPI bus is really done with it's
pump_transfers routine. Therefore this patch always tries
to read the voltage values through the SPI bus. And if that fails,
it tries to read the voltage values with the functions provided in
this patch.

This way of reading the voltage values will cause a lot of dev_err
output from the max1111 chip. This patch also remove this dev_err
line.